+++ packages/component-base/README.md
@@ -5,3 +5,5 @@ A set of mixins used by Vaadin components.
 ## License
 
 Apache License 2.0
+
+Vaadin collects development time usage statistics to improve this product. For details and to opt-out, see https://github.com/vaadin/vaadin-usage-statistics.
