+++ packages/component-base/index.js
@@ -1,5 +1,6 @@
 export { ActiveMixin } from './src/active-mixin.js';
 export { DisabledMixin } from './src/disabled-mixin.js';
+export { ElementMixin } from './src/vaadin-element-mixin.js';
 export { FocusMixin } from './src/focus-mixin.js';
 export { KeyboardMixin } from './src/keyboard-mixin.js';
 export { SlotMixin } from './src/slot-mixin.js';
