+++ packages/component-base/src/vaadin-dir-helper.js
