--- packages/vaadin-element-mixin/README.md
@@ -1,12 +1,0 @@
-# vaadin-element-mixin
-
-Provides common functionality for all Vaadin components.
-
-[![npm version](https://badgen.net/npm/v/@vaadin/vaadin-element-mixin)](https://www.npmjs.com/package/@vaadin/vaadin-element-mixin)
-[![Discord](https://img.shields.io/discord/732335336448852018?label=discord)](https://discord.gg/PHmkCKC)
-
-## License
-
-Apache License 2.0
-
-Vaadin collects development time usage statistics to improve this product. For details and to opt-out, see https://github.com/vaadin/vaadin-usage-statistics.
