--- packages/vaadin-element-mixin/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@vaadin/vaadin-element-mixin",
-  "version": "22.0.0-alpha6",
-  "description": "vaadin-element-mixin",
-  "main": "vaadin-element-mixin.js",
-  "module": "vaadin-element-mixin.js",
-  "repository": "vaadin/vaadin-element-mixin",
-  "keywords": [
-    "Vaadin",
-    "web-components",
-    "web-component",
-    "polymer"
-  ],
-  "author": "Vaadin Ltd",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/vaadin/vaadin-element-mixin/issues"
-  },
-  "homepage": "https://vaadin.com/components",
-  "files": [
-    "custom_typings",
-    "vaadin-*.d.ts",
-    "vaadin-*.js",
-    "templates.js"
-  ],
-  "dependencies": {
-    "@polymer/polymer": "^3.0.0",
-    "@vaadin/vaadin-development-mode-detector": "^2.0.0",
-    "@vaadin/vaadin-usage-statistics": "^2.1.0"
-  },
-  "devDependencies": {
-    "@esm-bundle/chai": "^4.3.4",
-    "@vaadin/testing-helpers": "^0.3.0",
-    "sinon": "^9.2.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
